+++ src/som/primitives/StringPrimitives.java
@@ -99,7 +99,8 @@ public class StringPrimitives extends Primitives {
 
         try {
           frame.push(universe.newString(self.getEmbeddedString().substring(
-              start.getEmbeddedInteger() - 1, end.getEmbeddedInteger())));
+              (int) start.getEmbeddedInteger() - 1,
+              (int) end.getEmbeddedInteger())));
         } catch (IndexOutOfBoundsException e) {
           frame.push(universe.newString(new java.lang.String(
               "Error - index out of bounds")));
